public/code/interest_rates/requirements.txt
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scipy==1.17.0
six==1.17.0
tzdata==2025.3
